Moving Client and Server to separate files to resolve recursive import
- Move MarionetteException to its own exceptions.py file
- Move Client class to client.py
- Move Server class to server.py
- Change imports in client.py, server.py, __init__.py to be relative

Fixes #47
