+++ marionette_tg/exceptions.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+
+class MarionetteException(Exception):
+    pass
